78587716/requirements.txt
python-dotenv==1.0.1
serpapi==0.1.5
